requirements.txt
fastapi~=0.115
uvicorn[standard]~=0.31
sqlalchemy~=2.0
psycopg2-binary~=2.9
pgvector~=0.3
pydantic~=2.9
pydantic-settings~=2.4
httpx~=0.28
requests~=2.32
python-dotenv~=1.0


